Bounties
Added a Bounty System for Player Killers. More players you kill, the larger the XP bounty on your head is.. When you die players can claim this bounty.

Planned:
Players with a bounty of 1000 XP+ Will have a distance from a player using the command to give a general idea where they are located.
(Assuming they are online)

Players with 3000+ XP bounty will have their coordinates shown in the /bounty command (Assuming they are online)

Players with 5000+ XP bounty will have all of the above facts and will be given a permanent glowing effect showing their location to nearby players
(Colours ranging from white to red depending on how high the bounty is)

Experience Gain:
Modified the Experience gain to allow for faster leveling..
I want it to be hard to level up, not impossible.