Oh The Bugs!
Plenty-O-Bugs

Realized that it is possible for a players data to not exist if they lose their lives at the same time as an event running.
This caused an error to spam the server console which although it doesnt affect anything it still looks horrible so figured i'd just quickly check it to make sure it is functional.