XRay!
I added an "XRay" feature into the plugin..

This is planned to be apart of a custom item as time goes on. It will have 3 tiers:
Tier 1 -> 10x10x10 block range
Tier 2 -> 15x15x15 block range
Tier 3 -> 20x20x20 block range

Notes:
This will lag the player, even more if they are around a lot of ores. But it'll be very very useful for mining and getting resources.

Current future plans:
Make it to where only the player using the item can see the XRay Effect.
Why? Because otherwise other players just passing by will know there is a player around them using this over powered item, this may cause targeted PKing. As well players may lag or even crash due to the amount of entities that could be spawned.