Please Kill Me Now!
Minecraft reflection is about as annoying as it gets!

But that being said, I have done a solid job at atleast knowing how to make it function at all.. It doesn't function great, but it still works!!!!!

So things to keep in mind for anyone wanting to read this over:
DONT!!!
You will probably have a stroke, lets be honest.. I was having a stroke while writing it. But it works! so its staying, its probably the worse possible way I could have made it, but that doesn't matter since it is currently 1:20 AM and I am not in the mood to try and clean up the code (not that it would be too hard).

So here is what was added:
The 'XRay' System now only shows for the player who will use the item! It will not show for any other players. This means people cant cause others to crash by spamming their xray ability.. The only people they will crash, is themself!

Something else.. Made a ReflectionUtility class to allow for some future reflection systems should I need them.. Hopefully I wont, because god this was a living hell to get functioning!